feat: types and moving mocks
