+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './menu.ts';
+export * from './review.ts';
+export * from './restaurant.ts';
