+++ src/types/menu.ts
@@ -1,0 +1,6 @@
+export type Menu = {
+  id: string;
+  name: string;
+  price: number;
+  ingredients: string[];
+};
