+++ src/types/restaurant.ts
@@ -1,0 +1,9 @@
+import type { Menu } from './menu.ts';
+import type { Review } from './review.ts';
+
+export type Restaurant = {
+  id: string;
+  name: string;
+  menu: Menu[];
+  reviews: Review[];
+};
