+++ src/types/review.ts
@@ -1,0 +1,6 @@
+export type Review = {
+  id: string;
+  user: string;
+  text: string;
+  rating: number;
+};
